requirements.txt
matplotlib
seaborn 
numpy
scikit-learn 
pillow
torch 
torchvision
